\name{ksample-package}
\alias{ksample-package}
\docType{package}
\title{K-sample tests}
\description{This package includes R ports of the Maximum Mean Discrepancy (MMD) two-sample test 
code of Gretton et al., and the Dynamic Slicing approach of Jiang et al.}
\details{
\tabular{ll}{
Package: \tab ksample\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-23\cr
License: \tab GPL-3\cr
}
}
\author{
MMD: Arthur Gretton, Heiko Strathmann.
DynSlicing: Bo Jiang, Chao Ye, and Jun S. Liu.
R port by Shachar Kaufman <shachar.kaufman@math.tau.ac.il>
}
\references{
Gretton, A., Sriperumbudur, B., Sejdinovic, D., Strathmann, H.,
Balakrishnan, S., Pontil, M., & Fukumizu, K.
Optimal kernel choice for large-scale two-sample tests. 
Advances in Neural Information Processing Systems, 2012.

Jiang, B., Ye, C. and Liu, J.S. 
Non-parametric K-sample tests via dynamic slicing. 
Journal of the American Statistical Association, 2014.
}
\seealso{
TODO.
}
\examples{
TODO.
}
